+++ CHANGES.md
@@ -41,6 +41,7 @@ To be released.
     [[#1164], [#1978]]
  -  Added `TxId.FromString()` static method.  [[#1978]]
  -  (Libplanet.Node) Added `UntypedTransaction` class.  [[#1974], [#1978]]
+ -  (Libplanet.Node) Added `UntypedBlock` class.  [[#1974], [#1978]]
 
 ### Behavioral changes